# bpo-47012: speed up iteration of bytes and bytearray