--- a/Misc/NEWS.d/next/Core and Builtins/2022-03-14-11-15-11.bpo-47012.5L6NoE.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-03-14-11-15-11.bpo-47012.5L6NoE.rst
@@ -1,0 +1,1 @@
+Speed up iteration of :class:`bytes` and :class:`bytearray` by 30%. Patch by Kumar Aditya.