Benchmark:

```py
from pyperf import Runner, perf_counter

def bench_bytes(loops, length):
    src = b'helloworld' * length
    t0 = perf_counter()
    for _ in range(loops):
        for i in src:
            pass
    return perf_counter() - t0

def bench_bytearray(loops, length):
    src = bytearray(b'hello' * length)
    t0 = perf_counter()
    for _ in range(loops):
        for i in src:
            pass
    return perf_counter() - t0

runner = Runner()
for n in [10_000, 100_000]:
    runner.bench_time_func(f"bytes {n}", bench_bytes, n)
    runner.bench_time_func(f"bytearray {n}", bench_bytearray, n)
```

Results:
```console
bytes 10000: Mean +- std dev: [base] 829 us +- 38 us -> [patch] 677 us +- 44 us: 1.23x faster
bytearray 10000: Mean +- std dev: [base] 523 us +- 34 us -> [patch] 360 us +- 19 us: 1.45x faster
bytes 100000: Mean +- std dev: [base] 8.33 ms +- 0.38 ms -> [patch] 6.89 ms +- 0.75 ms: 1.21x faster
bytearray 100000: Mean +- std dev: [base] 5.19 ms +- 0.23 ms -> [patch] 3.61 ms +- 0.23 ms: 1.44x faster

Geometric mean: 1.33x faster
```

<!-- issue-number: [bpo-47012](https://bugs.python.org/issue47012) -->
https://bugs.python.org/issue47012
<!-- /issue-number -->