@sweeneyde 2022-03-19
I agree with @animalize that it would be safest to include a preprocessor directive, whether that's with separate `#if _PY_NSMALLPOSINTS > 255`/`#else` code, or with a `#error` directive.

I'm not sure why anyone would compile with fewer than 256 cached small ints. [Here](https://github.com/python/cpython/commit/8e3b9f92835654943bb59d9658bb52e1b0f40a22), @vstinner added the code

```c
// _PyLong_GetZero() and _PyLong_GetOne() must always be available
#if _PY_NSMALLPOSINTS < 2
#  error "_PY_NSMALLPOSINTS must be greater than 1"
#endif
```

@vstinner, would there be any downside to requiring all of `(0, 1, ..., 255)` be in the small int cache?